Required changes and bug fixes
