+++ librz/arch/isa/amd29k/amd29k.c
@@ -22,7 +22,7 @@
 // Local registers
 #define AMD29K_IS_REG_LR(x) ((x) >= 128 && (x) < 256)
 #define AMD29K_REGNAME(x)   (AMD29K_IS_REG_GR(x) ? "gr" : "lr")
-#define AMD29K_LR(x)        (AMD29K_IS_REG_GR(x) ? (x) : (x)-127)
+#define AMD29K_LR(x)        (AMD29K_IS_REG_GR(x) ? (x) : (x) - 127)
 
 static void decode_ra_rb_rci(amd29k_instr_t *instruction, const ut8 *buffer) {
 	AMD29K_SET_VALUE(instruction, 0, buffer[1], AMD29K_TYPE_REG);
