+++ librz/arch/isa/avr/assembler.c
@@ -131,7 +131,7 @@
 			(rn) = abs - pc; \
 		} \
 		if ((rn) < 0) { \
-			(rn) = ~(-((rn)-1)); \
+			(rn) = ~(-((rn) - 1)); \
 		} else { \
 			(rn) -= 2; \
 		} \
