+++ librz/arch/isa/lh5801/lh5801.c
@@ -131,8 +131,8 @@ enum lh5801_insn_format {
 	LH5801_IFMT_RMODE_MASK = 3 << 10,
 };
 
-#define LH5801_IFMT_IMMS(f)  ((f)&LH5801_IFMT_IMM_MASK)
-#define LH5801_IFMT_RMODE(f) ((f)&LH5801_IFMT_RMODE_MASK)
+#define LH5801_IFMT_IMMS(f)  ((f) & LH5801_IFMT_IMM_MASK)
+#define LH5801_IFMT_RMODE(f) ((f) & LH5801_IFMT_RMODE_MASK)
 
 static bool lh5801_ifmt_fd_matches(enum lh5801_insn_format fmt, int fd) {
 	switch (fmt & LH5801_IFMT_FD_MASK) {
