+++ librz/arch/isa/v810/v810_disas.h
@@ -7,17 +7,17 @@
 #define V810_INSTR_MAXLEN 24
 
 #define OPCODE(instr) (((instr) >> 10) & 0x3F)
-#define REG1(instr)   ((instr)&0x1F)
+#define REG1(instr)   ((instr) & 0x1F)
 #define REG2(instr)   (((instr) >> 5) & 0x1F)
 #define IMM5(instr)   REG1((instr))
 #define COND(instr)   (((instr) >> 9) & 0xF)
 
-#define SIGN_EXT_T5(imm)  (((imm)&0x10) ? (imm) | 0xE0 : (imm))
-#define SIGN_EXT_T9(imm)  (((imm)&0x100) ? (imm) | 0xFFFFFE00 : (imm))
-#define SIGN_EXT_T26(imm) (((imm)&0x2000000) ? (imm) | 0xFC000000 : (imm))
+#define SIGN_EXT_T5(imm)  (((imm) & 0x10) ? (imm) | 0xE0 : (imm))
+#define SIGN_EXT_T9(imm)  (((imm) & 0x100) ? (imm) | 0xFFFFFE00 : (imm))
+#define SIGN_EXT_T26(imm) (((imm) & 0x2000000) ? (imm) | 0xFC000000 : (imm))
 
-#define DISP9(word1)         SIGN_EXT_T9((word1)&0x1FE)
-#define DISP26(word1, word2) SIGN_EXT_T26((((word1)&0x3FF) << 16) | (word2))
+#define DISP9(word1)         SIGN_EXT_T9((word1) & 0x1FE)
+#define DISP26(word1, word2) SIGN_EXT_T26((((word1) & 0x3FF) << 16) | (word2))
 
 enum v810_cmd_opcodes {
 	V810_MOV = 0x0,
