+++ librz/bin/format/elf/glibc_elf.h
@@ -418,7 +418,7 @@ typedef struct
 #define SHN_LORESERVE 0xff00 /* Start of reserved indices */
 #define SHN_LOPROC    0xff00 /* Start of processor-specific */
 #define SHN_BEFORE    0xff00 /* Order section before all others \
-				(Solaris).  */
+			     (Solaris).  */
 #define SHN_AFTER 0xff01 /* Order section after all others \
 			    (Solaris).  */
 #define SHN_HIPROC    0xff1f /* End of processor-specific */
@@ -485,7 +485,7 @@ typedef struct
 #define SHF_MASKOS     0x0ff00000 /* OS-specific.  */
 #define SHF_MASKPROC   0xf0000000 /* Processor-specific */
 #define SHF_ORDERED    (1 << 30) /* Special ordering requirement \
-				    (Solaris).  */
+				 (Solaris).  */
 #define SHF_EXCLUDE (1U << 31) /* Section is excluded unless \
 				  referenced or allocated (Solaris).*/
 
@@ -573,8 +573,8 @@ typedef struct
 /* How to extract and insert information held in the st_info field.  */
 
 #define ELF32_ST_BIND(val)        (((unsigned char)(val)) >> 4)
-#define ELF32_ST_TYPE(val)        ((val)&0xf)
-#define ELF32_ST_INFO(bind, type) (((bind) << 4) + ((type)&0xf))
+#define ELF32_ST_TYPE(val)        ((val) & 0xf)
+#define ELF32_ST_INFO(bind, type) (((bind) << 4) + ((type) & 0xf))
 
 /* Both Elf32_Sym and Elf64_Sym use the same one-byte st_info field.  */
 #define ELF64_ST_BIND(val)        ELF32_ST_BIND(val)
@@ -617,7 +617,7 @@ typedef struct
 
 /* How to extract and insert information held in the st_other field.  */
 
-#define ELF32_ST_VISIBILITY(o) ((o)&0x03)
+#define ELF32_ST_VISIBILITY(o) ((o) & 0x03)
 
 /* For ELF64 the definitions are the same.  */
 #define ELF64_ST_VISIBILITY(o) ELF32_ST_VISIBILITY(o)
@@ -666,11 +666,11 @@ typedef struct
 /* How to extract and insert information held in the rz_info field.  */
 
 #define ELF32_R_SYM(val)        ((val) >> 8)
-#define ELF32_R_TYPE(val)       ((val)&0xff)
-#define ELF32_R_INFO(sym, type) (((sym) << 8) + ((type)&0xff))
+#define ELF32_R_TYPE(val)       ((val) & 0xff)
+#define ELF32_R_INFO(sym, type) (((sym) << 8) + ((type) & 0xff))
 
 #define ELF64_R_SYM(i)          ((i) >> 32)
-#define ELF64_R_TYPE(i)         ((i)&0xffffffff)
+#define ELF64_R_TYPE(i)         ((i) & 0xffffffff)
 #define ELF64_R_INFO(sym, type) ((((Elf64_Xword)(sym)) << 32) + (type))
 
 /* Program segment header.  */
@@ -741,7 +741,7 @@ typedef struct
 
 #define NT_PRSTATUS 1 /* Contains copy of prstatus struct */
 #define NT_PRFPREG  2 /* Contains copy of fpregset \
-			 struct.  */
+			struct.  */
 #define NT_FPREGSET   2 /* Contains copy of fpregset struct */
 #define NT_PRPSINFO   3 /* Contains copy of prpsinfo struct */
 #define NT_PRXREG     4 /* Contains copy of prxregset struct */
@@ -758,7 +758,7 @@ typedef struct
 #define NT_LWPSINFO   17 /* Contains copy of lwpinfo struct */
 #define NT_PRFPXREG   20 /* Contains copy of fprxregset struct */
 #define NT_SIGINFO    0x53494749 /* Contains copy of siginfo_t, \
-				    size might increase */
+				 size might increase */
 #define NT_FILE 0x46494c45 /* Contains information about mapped \
 			      files */
 #define NT_PRXFPREG    0x46e62b7f /* Contains copy of user_fxsr_struct */
@@ -796,11 +796,11 @@ typedef struct
 #define NT_S390_SYSTEM_CALL 0x307 /* s390 system call restart data */
 #define NT_S390_TDB         0x308 /* s390 transaction diagnostic block */
 #define NT_S390_VXRS_LOW    0x309 /* s390 vector registers 0-15 \
-				     upper half.  */
+				  upper half.  */
 #define NT_S390_VXRS_HIGH 0x30a /* s390 vector registers 16-31.  */
 #define NT_S390_GS_CB     0x30b /* s390 guarded storage registers.  */
 #define NT_S390_GS_BC     0x30c /* s390 guarded storage \
-				   broadcast control block.  */
+			       broadcast control block.  */
 #define NT_S390_RI_CB      0x30d /* s390 runtime instrumentation.  */
 #define NT_ARM_VFP         0x400 /* ARM VFP/NEON registers */
 #define NT_ARM_TLS         0x401 /* ARM TLS register */
@@ -808,7 +808,7 @@ typedef struct
 #define NT_ARM_HW_WATCH    0x403 /* ARM hardware watchpoint registers */
 #define NT_ARM_SYSTEM_CALL 0x404 /* ARM system call number */
 #define NT_ARM_SVE         0x405 /* ARM Scalable Vector Extension \
-				    registers */
+			    registers */
 #define NT_ARM_PAC_MASK 0x406 /* ARM pointer authentication \
 				 code masks.  */
 #define NT_ARM_PACA_KEYS 0x407 /* ARM pointer authentication \
@@ -901,7 +901,7 @@ typedef struct
 #define DT_MOVESZ         0x6ffffdfb
 #define DT_FEATURE_1      0x6ffffdfc /* Feature selection (DTF_*).  */
 #define DT_POSFLAG_1      0x6ffffdfd /* Flags for DT_* entries, effecting \
-					the following DT_* entry.  */
+				   the following DT_* entry.  */
 #define DT_SYMINSZ        0x6ffffdfe /* Size of syminfo table (in bytes) */
 #define DT_SYMINENT       0x6ffffdff /* Entry size of syminfo */
 #define DT_VALRNGHI       0x6ffffdff
@@ -939,10 +939,10 @@ typedef struct
 /* These were chosen by Sun.  */
 #define DT_FLAGS_1 0x6ffffffb /* State flags, see DF_1_* below.  */
 #define DT_VERDEF  0x6ffffffc /* Address of version definition \
-				 table */
+				table */
 #define DT_VERDEFNUM 0x6ffffffd /* Number of version definitions */
 #define DT_VERNEED   0x6ffffffe /* Address of table with needed \
-				   versions */
+				 versions */
 #define DT_VERNEEDNUM         0x6fffffff /* Number of needed versions */
 #define DT_VERSIONTAGIDX(tag) (DT_VERNEEDNUM - (tag)) /* Reverse order! */
 #define DT_VERSIONTAGNUM      16
@@ -1168,7 +1168,7 @@ typedef struct
 /* Some more special a_type values describing the hardware.  */
 #define AT_PLATFORM 15 /* String identifying platform.  */
 #define AT_HWCAP    16 /* Machine-dependent hints about \
-			  processor capabilities.  */
+		       processor capabilities.  */
 
 /* This entry gives some information about the FPU initialization
    performed by the kernel.  */
@@ -1418,7 +1418,7 @@ typedef struct
 #define RZ_68K_TLS_IE16  35 /* 16 bit GOT offset for IE */
 #define RZ_68K_TLS_IE8   36 /* 8 bit GOT offset for IE */
 #define RZ_68K_TLS_LE32  37 /* 32 bit offset relative to \
-			      static TLS block */
+			     static TLS block */
 #define RZ_68K_TLS_LE16 38 /* 16 bit offset relative to \
 			     static TLS block */
 #define RZ_68K_TLS_LE8 39 /* 8 bit offset relative to \
@@ -1447,7 +1447,7 @@ typedef struct
 #define RZ_386_32PLT     11
 #define RZ_386_TLS_TPOFF 14 /* Offset in static TLS block */
 #define RZ_386_TLS_IE    15 /* Address of GOT entry for static TLS \
-			       block offset */
+			    block offset */
 #define RZ_386_TLS_GOTIE 16 /* GOT entry for static TLS block \
 		       offset */
 #define RZ_386_TLS_LE 17 /* Offset relative to static TLS \
@@ -1475,7 +1475,7 @@ typedef struct
 #define RZ_386_TLS_LDM_POP 31 /* Tag for popl in LDM TLS code */
 #define RZ_386_TLS_LDO_32  32 /* Offset relative to TLS block */
 #define RZ_386_TLS_IE_32   33 /* GOT entry for negated static TLS \
-			 block offset */
+		       block offset */
 #define RZ_386_TLS_LE_32 34 /* Negated offset relative to static \
 		       TLS block */
 #define RZ_386_TLS_DTPMOD32  35 /* ID of module containing symbol */
@@ -1492,7 +1492,7 @@ typedef struct
 			      offset for the symbol.  */
 #define RZ_386_IRELATIVE 42 /* Adjust indirectly by program base */
 #define RZ_386_GOT32X    43 /* Load from 32 bit GOT entry, \
-			       relaxable. */
+			    relaxable. */
 /* Keep this the last entry.  */
 #define RZ_386_NUM 44
 
@@ -2649,7 +2649,7 @@ enum {
 #define STO_PPC64_LOCAL_BIT  5
 #define STO_PPC64_LOCAL_MASK (7 << STO_PPC64_LOCAL_BIT)
 #define PPC64_LOCAL_ENTRY_OFFSET(other) \
-	(((1 << (((other)&STO_PPC64_LOCAL_MASK) >> STO_PPC64_LOCAL_BIT)) >> 2) << 2)
+	(((1 << (((other) & STO_PPC64_LOCAL_MASK) >> STO_PPC64_LOCAL_BIT)) >> 2) << 2)
 
 /* ARM specific declarations */
 
@@ -2681,7 +2681,7 @@ enum {
 #define EF_ARM_BE8 0x00800000
 #define EF_ARM_LE8 0x00400000
 
-#define EF_ARM_EABI_VERSION(flags) ((flags)&EF_ARM_EABIMASK)
+#define EF_ARM_EABI_VERSION(flags) ((flags) & EF_ARM_EABIMASK)
 #define EF_ARM_EABI_UNKNOWN        0x00000000
 #define EF_ARM_EABI_VER1           0x01000000
 #define EF_ARM_EABI_VER2           0x02000000
@@ -2696,7 +2696,7 @@ enum {
 /* ARM-specific values for sh_flags */
 #define SHF_ARM_ENTRYSECT 0x10000000 /* Section contains an entry point */
 #define SHF_ARM_COMDEF    0x80000000 /* Section may be multiply defined \
-					in the input to a link step.  */
+				     in the input to a link step.  */
 
 /* ARM-specific program header flags */
 #define PF_ARM_SB 0x10000000 /* Segment contains the location \
@@ -2874,7 +2874,7 @@ enum {
 #define RZ_ARM_SBREL32  9
 #define RZ_ARM_THM_PC22 10 /* PC relative 24 bit (Thumb32 BL).  */
 #define RZ_ARM_THM_PC8  11 /* PC relative & 0x3FC \
-			      (Thumb16 LDR, ADD, ADR).  */
+			     (Thumb16 LDR, ADD, ADR).  */
 #define RZ_ARM_AMP_VCALL9   12
 #define RZ_ARM_SWI24        13 /* Obsolete static relocation.  */
 #define RZ_ARM_TLS_DESC     13 /* Dynamic relocation.  */
@@ -2894,7 +2894,7 @@ enum {
 #define RZ_ARM_PLT32        27 /* Deprecated, 32 bit PLT address.  */
 #define RZ_ARM_CALL         28 /* PC relative 24 bit (BL, BLX).  */
 #define RZ_ARM_JUMP24       29 /* PC relative 24 bit \
-				  (B, BL<cond>).  */
+			    (B, BL<cond>).  */
 #define RZ_ARM_THM_JUMP24      30 /* PC relative 24 bit (Thumb32 B.W).  */
 #define RZ_ARM_BASE_ABS        31 /* Adjust by program base.  */
 #define RZ_ARM_ALU_PCREL_7_0   32 /* Obsolete.  */
@@ -2914,7 +2914,7 @@ enum {
 #define RZ_ARM_MOVT_PREL       46 /* PC relative (MOVT).  */
 #define RZ_ARM_THM_MOVW_ABS_NC 47 /* Direct 16 bit (Thumb32 MOVW).  */
 #define RZ_ARM_THM_MOVT_ABS    48 /* Direct high 16 bit \
-				     (Thumb32 MOVT).  */
+				  (Thumb32 MOVT).  */
 #define RZ_ARM_THM_MOVW_PREL_NC 49 /* PC relative 16 bit \
 				      (Thumb32 MOVW).  */
 #define RZ_ARM_THM_MOVT_PREL 50 /* PC relative high 16 bit \
@@ -2937,7 +2937,7 @@ enum {
 #define RZ_ARM_LDR_PC_G1    62 /* PC relative (LDR,STR,LDRB,STRB).  */
 #define RZ_ARM_LDR_PC_G2    63 /* PC relative (LDR,STR,LDRB,STRB).  */
 #define RZ_ARM_LDRS_PC_G0   64 /* PC relative (STR{D,H}, \
-				  LDR{D,SB,H,SH}).  */
+				LDR{D,SB,H,SH}).  */
 #define RZ_ARM_LDRS_PC_G1 65 /* PC relative (STR{D,H}, \
 				LDR{D,SB,H,SH}).  */
 #define RZ_ARM_LDRS_PC_G2 66 /* PC relative (STR{D,H}, \
@@ -2951,7 +2951,7 @@ enum {
 #define RZ_ARM_ALU_SB_G1    73 /* Program base relative (ADD,SUB).  */
 #define RZ_ARM_ALU_SB_G2    74 /* Program base relative (ADD,SUB).  */
 #define RZ_ARM_LDR_SB_G0    75 /* Program base relative (LDR, \
-			  STR, LDRB, STRB).  */
+		       STR, LDRB, STRB).  */
 #define RZ_ARM_LDR_SB_G1 76 /* Program base relative \
 		       (LDR, STR, LDRB, STRB).  */
 #define RZ_ARM_LDR_SB_G2 77 /* Program base relative \
@@ -2985,7 +2985,7 @@ enum {
 #define RZ_ARM_GOT_ABS      95 /* GOT entry.  */
 #define RZ_ARM_GOT_PREL     96 /* PC relative GOT entry.  */
 #define RZ_ARM_GOT_BREL12   97 /* GOT entry relative to GOT \
-				  origin (LDR).  */
+				origin (LDR).  */
 #define RZ_ARM_GOTOFF12 98 /* 12 bit, GOT entry relative \
 			      to GOT origin (LDR, STR).  */
 #define RZ_ARM_GOTRELAX      99
@@ -2993,7 +2993,7 @@ enum {
 #define RZ_ARM_GNU_VTINHERIT 101
 #define RZ_ARM_THM_PC11      102 /* PC relative & 0xFFE (Thumb16 B).  */
 #define RZ_ARM_THM_PC9       103 /* PC relative & 0x1FE \
-				    (Thumb16 B/B<cond>).  */
+			      (Thumb16 B/B<cond>).  */
 #define RZ_ARM_TLS_GD32 104 /* PC-rel 32 bit for global dynamic \
 			       thread local data */
 #define RZ_ARM_TLS_LDM32 105 /* PC-rel 32 bit for local dynamic \
@@ -3015,7 +3015,7 @@ enum {
 #define RZ_ARM_THM_TLS_DESCSEQ16 129
 #define RZ_ARM_THM_TLS_DESCSEQ32 130
 #define RZ_ARM_THM_GOT_BREL12    131 /* GOT entry relative to GOT \
-					origin, 12 bit (Thumb32 LDR).  */
+				     origin, 12 bit (Thumb32 LDR).  */
 #define RZ_ARM_IRELATIVE 160
 #define RZ_ARM_RXPC25    249
 #define RZ_ARM_RSBREL32  250
@@ -3362,7 +3362,7 @@ enum {
 #define RZ_390_TLS_DTPMOD 54 /* ID of module containing symbol.  */
 #define RZ_390_TLS_DTPOFF 55 /* Offset in TLS block.	 */
 #define RZ_390_TLS_TPOFF  56 /* Negated offset in static TLS \
-			block.  */
+		       block.  */
 #define RZ_390_20          57 /* Direct 20 bit.  */
 #define RZ_390_GOT20       58 /* 20 bit GOT offset.  */
 #define RZ_390_GOTPLT20    59 /* 20 bit offset to jump slot.  */
@@ -3407,7 +3407,7 @@ enum {
 #define RZ_X86_64_JUMP_SLOT 7 /* Create PLT entry */
 #define RZ_X86_64_RELATIVE  8 /* Adjust by program base */
 #define RZ_X86_64_GOTPCREL  9 /* 32 bit signed PC relative \
-				 offset to GOT */
+				offset to GOT */
 #define RZ_X86_64_32       10 /* Direct 32 bit zero extended */
 #define RZ_X86_64_32S      11 /* Direct 32 bit sign extended */
 #define RZ_X86_64_16       12 /* Direct 16 bit zero extended */
@@ -3418,7 +3418,7 @@ enum {
 #define RZ_X86_64_DTPOFF64 17 /* Offset in module's TLS block */
 #define RZ_X86_64_TPOFF64  18 /* Offset in initial TLS block */
 #define RZ_X86_64_TLSGD    19 /* 32 bit signed PC relative offset \
-				 to two GOT entries for GD symbol */
+			      to two GOT entries for GD symbol */
 #define RZ_X86_64_TLSLD 20 /* 32 bit signed PC relative offset \
 			      to two GOT entries for LD symbol */
 #define RZ_X86_64_DTPOFF32 21 /* Offset in TLS block */
@@ -3428,7 +3428,7 @@ enum {
 #define RZ_X86_64_PC64     24 /* PC relative 64 bit */
 #define RZ_X86_64_GOTOFF64 25 /* 64 bit offset to GOT */
 #define RZ_X86_64_GOTPC32  26 /* 32 bit signed pc relative \
-				 offset to GOT */
+				offset to GOT */
 #define RZ_X86_64_GOT64      27 /* 64-bit GOT entry offset */
 #define RZ_X86_64_GOTPCREL64 28 /* 64-bit PC relative offset \
 				   to GOT entry */
@@ -3440,7 +3440,7 @@ enum {
 #define RZ_X86_64_SIZE64          33 /* Size of symbol plus 64-bit addend */
 #define RZ_X86_64_GOTPC32_TLSDESC 34 /* GOT offset for TLS descriptor.  */
 #define RZ_X86_64_TLSDESC_CALL    35 /* Marker for call through TLS \
-					descriptor.  */
+				     descriptor.  */
 #define RZ_X86_64_TLSDESC    36 /* TLS descriptor.  */
 #define RZ_X86_64_IRELATIVE  37 /* Adjust indirectly by program base */
 #define RZ_X86_64_RELATIVE64 38 /* 64-bit adjust by program base */
@@ -3486,7 +3486,7 @@ enum {
 #define RZ_MN10300_TLS_LD        25 /* 32-bit offset for local dynamic.  */
 #define RZ_MN10300_TLS_LDO       26 /* Module-relative offset.  */
 #define RZ_MN10300_TLS_GOTIE     27 /* GOT offset for static TLS block \
-				       offset.  */
+				   offset.  */
 #define RZ_MN10300_TLS_IE 28 /* GOT address for static TLS block \
 				offset.  */
 #define RZ_MN10300_TLS_LE 29 /* Offset relative to static TLS \
@@ -3495,7 +3495,7 @@ enum {
 #define RZ_MN10300_TLS_DTPOFF 31 /* Offset in module TLS block.  */
 #define RZ_MN10300_TLS_TPOFF  32 /* Offset in static TLS block.  */
 #define RZ_MN10300_SYM_DIFF   33 /* Adjustment for next reloc as needed \
-				    by linker relaxation.  */
+				  by linker relaxation.  */
 #define RZ_MN10300_ALIGN 34 /* Alignment requirement for linker \
 		       relaxation.  */
 #define RZ_MN10300_NUM 35
@@ -3613,7 +3613,7 @@ enum {
 #define RZ_NIOS2_CJMP          19 /* Conditional branch.  */
 #define RZ_NIOS2_CALLR         20 /* Indirect call through register.  */
 #define RZ_NIOS2_ALIGN         21 /* Alignment requirement for \
-				     linker relaxation.  */
+			     linker relaxation.  */
 #define RZ_NIOS2_GOT16       22 /* 16 bit GOT entry.  */
 #define RZ_NIOS2_CALL16      23 /* 16 bit GOT entry for function.  */
 #define RZ_NIOS2_GOTOFF_LO   24 /* %lo of offset to GOT pointer.  */
