+++ librz/bin/p/bin_avr.c
@@ -7,15 +7,15 @@
 #define CHECK4INSTR(b, instr, size) \
 	if (!instr(b, 0) || \
 		!instr((b), (size)) || \
-		!instr((b), (size)*2) || \
-		!instr((b), (size)*3)) { \
+		!instr((b), (size) * 2) || \
+		!instr((b), (size) * 3)) { \
 		return false; \
 	}
 
 #define CHECK3INSTR(b, instr, size) \
 	if (!instr((b), (size)) || \
-		!instr((b), (size)*2) || \
-		!instr((b), (size)*3)) { \
+		!instr((b), (size) * 2) || \
+		!instr((b), (size) * 3)) { \
 		return false; \
 	}
