+++ librz/debug/p/native/android_x86_64.c
@@ -1,3 +1,6 @@
+// SPDX-FileCopyrightText: 2009-2019 pancake <pancake@nopcode.org>
+// SPDX-License-Identifier: LGPL-3.0-only
+
 #include <errno.h>
 #if !defined(__HAIKU__) && !defined(__sun)
 #include <sys/ptrace.h>
@@ -5,8 +8,13 @@
 #include <sys/wait.h>
 #include <signal.h>
 
-#include "native/bsd/bsd_debug.h"
-#include "native/procfs.h"
+#include <sys/mman.h>
+#include "linux/linux_debug.h"
+#include "procfs.h"
+#define WAIT_ANY -1
+#ifndef WIFCONTINUED
+#define WIFCONTINUED(s) ((s) == 0xffff)
+#endif
 
 #ifdef __WALL
 #define WAITPID_FLAGS __WALL
@@ -22,55 +30,85 @@
 #define PROC_PERM_SZ        5
 #define PROC_UNKSTR_SZ      128
 
+#if WAIT_ON_ALL_CHILDREN
 static int rz_debug_handle_signals(RzDebug *dbg) {
 	eprintf("Warning: signal handling is not supported on this platform\n");
 	return 0;
 }
+#endif
 
-#include "native/reg.c"
+static char *rz_debug_native_reg_profile(RzDebug *dbg) {
+	return linux_reg_profile(dbg);
+}
+
+#include "reg.c"
 
 static bool rz_debug_native_step(RzDebug *dbg) {
-	int ret = ptrace(PT_STEP, dbg->pid, (caddr_t)1, 0);
-	if (ret != 0) {
-		perror("native-singlestep");
-		return false;
-	}
-	return true;
+	return linux_step(dbg);
 }
 
 static int rz_debug_native_attach(RzDebug *dbg, int pid) {
-	int ret = ptrace(PTRACE_ATTACH, pid, 0, 0);
-	if (ret != -1) {
-		eprintf("Trying to attach to %d\n", pid);
-		perror("ptrace (PT_ATTACH)");
-	}
-	return pid;
+	return linux_attach(dbg, pid);
 }
 
 static int rz_debug_native_detach(RzDebug *dbg, int pid) {
-	return ptrace(PT_DETACH, pid, NULL, 0);
+	return rz_debug_ptrace(dbg, PTRACE_DETACH, pid, NULL, (rz_ptrace_data_t)(size_t)0);
+}
+
+static int rz_debug_native_select(RzDebug *dbg, int pid, int tid) {
+	return linux_select(dbg, pid, tid);
 }
 
 static int rz_debug_native_continue_syscall(RzDebug *dbg, int pid, int num) {
-	ut64 pc = rz_debug_reg_get(dbg, "PC");
-	errno = 0;
-	return ptrace(PTRACE_SYSCALL, pid, (void *)(size_t)pc, 0) == 0;
+	linux_set_options(dbg, pid);
+	return rz_debug_ptrace(dbg, PTRACE_SYSCALL, pid, 0, 0);
+}
+
+static void interrupt_process(RzDebug *dbg) {
+	rz_debug_kill(dbg, dbg->pid, dbg->tid, SIGINT);
+	rz_cons_break_pop();
 }
 
 static int rz_debug_native_stop(RzDebug *dbg) {
-	return 0;
+	return linux_stop_threads(dbg, dbg->reason.tid);
 }
 
 static int rz_debug_native_continue(RzDebug *dbg, int pid, int tid, int sig) {
-	void *data = (void *)(size_t)((sig != -1) ? sig : dbg->reason.signum);
-	ut64 pc = rz_debug_reg_get(dbg, "PC");
-	return ptrace(PTRACE_CONT, pid, (void *)(size_t)pc, (int)(size_t)data) == 0;
+	int contsig = dbg->reason.signum;
+	int ret = -1;
+
+	if (sig != -1) {
+		contsig = sig;
+	}
+	/* SIGINT handler for attached processes: dbg.consbreak (disabled by default) */
+	if (dbg->consbreak) {
+		rz_cons_break_push((RzConsBreak)interrupt_process, dbg);
+	}
+
+	if (dbg->continue_all_threads && dbg->n_threads && dbg->threads) {
+		RzDebugPid *th;
+		RzListIter *it;
+		rz_list_foreach (dbg->threads, it, th) {
+			ret = rz_debug_ptrace(dbg, PTRACE_CONT, th->pid, 0, 0);
+			if (ret) {
+				eprintf("Error: (%d) is running or dead.\n", th->pid);
+			}
+		}
+	} else {
+		ret = rz_debug_ptrace(dbg, PTRACE_CONT, tid, NULL, (rz_ptrace_data_t)(size_t)contsig);
+		if (ret) {
+			rz_sys_perror("PTRACE_CONT");
+		}
+	}
+	// return ret >= 0 ? tid : false;
+	return tid;
 }
 
 static RzDebugInfo *rz_debug_native_info(RzDebug *dbg, const char *arg) {
-	return bsd_info(dbg, arg);
+	return linux_info(dbg, arg);
 }
 
+#ifdef WAIT_ON_ALL_CHILDREN
 static RzDebugReasonType rz_debug_native_wait(RzDebug *dbg, int pid) {
 	RzDebugReasonType reason = RZ_DEBUG_REASON_UNKNOWN;
 
@@ -79,74 +117,100 @@ static RzDebugReasonType rz_debug_native_wait(RzDebug *dbg, int pid) {
 		return RZ_DEBUG_REASON_ERROR;
 	}
 	int status = -1;
-#ifdef WAIT_ON_ALL_CHILDREN
+	// XXX: this is blocking, ^C will be ignored
 	int ret = waitpid(-1, &status, WAITPID_FLAGS);
-#else
-	int ret = waitpid(-1, &status, 0);
-	if (ret != -1) {
-		reason = RZ_DEBUG_REASON_TRAP;
-	}
-#endif
 	if (ret == -1) {
 		rz_sys_perror("waitpid");
 		return RZ_DEBUG_REASON_ERROR;
 	}
 
 	// eprintf ("rz_debug_native_wait: status=%d (0x%x) (return=%d)\n", status, status, ret);
+
+	if (ret != pid) {
+		reason = RZ_DEBUG_REASON_NEW_PID;
+		eprintf("switching to pid %d\n", ret);
+		rz_debug_select(dbg, ret, ret);
+	}
+
+	// TODO: switch status and handle reasons here
+	// FIXME: Remove linux handling from this function?
+#if defined(PT_GETEVENTMSG)
+	reason = linux_ptrace_event(dbg, pid, status, true);
+#endif
+
+	/* propagate errors */
 	if (reason == RZ_DEBUG_REASON_ERROR) {
 		return reason;
 	}
+
 	/* we don't know what to do yet, let's try harder to figure it out. */
 	if (reason == RZ_DEBUG_REASON_UNKNOWN) {
-	if (WIFEXITED(status)) {
-		eprintf("child exited with status %d\n", WEXITSTATUS(status));
-		reason = RZ_DEBUG_REASON_DEAD;
-	} else if (WIFSIGNALED(status)) {
-		eprintf("child received signal %d\n", WTERMSIG(status));
-		reason = RZ_DEBUG_REASON_SIGNAL;
-	} else if (WIFSTOPPED(status)) {
-		if (WSTOPSIG(status) != SIGTRAP &&
-			WSTOPSIG(status) != SIGSTOP) {
-			eprintf("Child stopped with signal %d\n", WSTOPSIG(status));
-		}
+		if (WIFEXITED(status)) {
+			eprintf("child exited with status %d\n", WEXITSTATUS(status));
+			reason = RZ_DEBUG_REASON_DEAD;
+		} else if (WIFSIGNALED(status)) {
+			eprintf("child received signal %d\n", WTERMSIG(status));
+			reason = RZ_DEBUG_REASON_SIGNAL;
+		} else if (WIFSTOPPED(status)) {
+			if (WSTOPSIG(status) != SIGTRAP &&
+				WSTOPSIG(status) != SIGSTOP) {
+				eprintf("Child stopped with signal %d\n", WSTOPSIG(status));
+			}
 
-		/* the ptrace documentation says GETSIGINFO is only necessary for
-		 * differentiating the various stops.
-		 *
-		 * this might modify dbg->reason.signum
-		 */
-		reason = RZ_DEBUG_REASON_BREAKPOINT;
+			/* the ptrace documentation says GETSIGINFO is only necessary for
+			 * differentiating the various stops.
+			 *
+			 * this might modify dbg->reason.signum
+			 */
+			if (rz_debug_handle_signals(dbg) != 0) {
+				return RZ_DEBUG_REASON_ERROR;
+			}
+			reason = dbg->reason.type;
 #ifdef WIFCONTINUED
-	} else if (WIFCONTINUED(status)) {
-		eprintf("child continued...\n");
-		reason = RZ_DEBUG_REASON_NONE;
+		} else if (WIFCONTINUED(status)) {
+			eprintf("child continued...\n");
+			reason = RZ_DEBUG_REASON_NONE;
 #endif
-	} else if (status == 1) {
-		/* XXX(jjd): does this actually happen? */
-		eprintf("debugger is dead with status 1!\n");
-		reason = RZ_DEBUG_REASON_DEAD;
-	} else if (status == 0) {
-		/* XXX(jjd): does this actually happen? */
-		eprintf("debugger is dead with status 0\n");
-		reason = RZ_DEBUG_REASON_DEAD;
-	} else {
-		if (ret != pid) {
-			reason = RZ_DEBUG_REASON_NEW_PID;
+		} else if (status == 1) {
+			/* XXX(jjd): does this actually happen? */
+			eprintf("debugger is dead with status 1!\n");
+			reason = RZ_DEBUG_REASON_DEAD;
+		} else if (status == 0) {
+			/* XXX(jjd): does this actually happen? */
+			eprintf("debugger is dead with status 0\n");
+			reason = RZ_DEBUG_REASON_DEAD;
 		} else {
-			/* ugh. still don't know :-/ */
-			eprintf("returning from wait without knowing why...\n");
+			if (ret != pid) {
+				reason = RZ_DEBUG_REASON_NEW_PID;
+			} else {
+				/* ugh. still don't know :-/ */
+				eprintf("returning from wait without knowing why...\n");
+			}
 		}
 	}
+
+	/* if we still don't know what to do, we have a problem... */
+	if (reason == RZ_DEBUG_REASON_UNKNOWN) {
+		eprintf("%s: no idea what happened...\n", __func__);
+		reason = RZ_DEBUG_REASON_ERROR;
+	}
+	dbg->reason.tid = pid;
+	dbg->reason.type = reason;
+	return reason;
 }
-/* if we still don't know what to do, we have a problem... */
-if (reason == RZ_DEBUG_REASON_UNKNOWN) {
-	eprintf("%s: no idea what happened...\n", __func__);
-	reason = RZ_DEBUG_REASON_ERROR;
+#else
+static RzDebugReasonType rz_debug_native_wait(RzDebug *dbg, int pid) {
+	RzDebugReasonType reason = RZ_DEBUG_REASON_UNKNOWN;
+	if (pid == -1) {
+		eprintf("ERROR: rz_debug_native_wait called with pid -1\n");
+		return RZ_DEBUG_REASON_ERROR;
+	}
+
+	reason = linux_dbg_wait(dbg, dbg->tid);
+	dbg->reason.type = reason;
+	return reason;
 }
-dbg->reason.tid = pid;
-dbg->reason.type = reason;
-return reason;
-}
+#endif
 
 #undef MAXPID
 #define MAXPID 99999
@@ -156,7 +220,7 @@ static RzList /*<RzDebugPid *>*/ *rz_debug_native_pids(RzDebug *dbg, int pid) {
 	if (!list) {
 		return NULL;
 	}
-	return bsd_pid_list(dbg, pid, list);
+	return linux_pid_list(pid, list);
 }
 
 RZ_API RZ_OWN RzList /*<RzDebugPid *>*/ *rz_debug_native_threads(RzDebug *dbg, int pid) {
@@ -165,84 +229,161 @@ RZ_API RZ_OWN RzList /*<RzDebugPid *>*/ *rz_debug_native_threads(RzDebug *dbg, i
 		eprintf("No list?\n");
 		return NULL;
 	}
-	return bsd_thread_list(dbg, pid, list);
+	return linux_thread_list(dbg, pid, list);
 }
 
 RZ_API ut64 rz_debug_get_tls(RZ_NONNULL RzDebug *dbg, int tid) {
 	rz_return_val_if_fail(dbg, 0);
-	return 0;
-}
-
-static int bsd_reg_read(RzDebug *dbg, int type, ut8 *buf, int size) {
-	int showfpu = false;
-	int pid = dbg->pid;
-	int ret;
-	if (type < -1) {
-		showfpu = true; // hack for debugging
-		type = -type;
-	}
-	switch (type) {
-	case RZ_REG_TYPE_DRX:
-		return true;
-		break;
-	case RZ_REG_TYPE_FPU:
-	case RZ_REG_TYPE_MMX:
-	case RZ_REG_TYPE_XMM:
-		break;
-	case RZ_REG_TYPE_SEG:
-	case RZ_REG_TYPE_FLG:
-	case RZ_REG_TYPE_GPR: {
-		RZ_DEBUG_REG_T regs;
-		memset(&regs, 0, sizeof(regs));
-		memset(buf, 0, size);
-#warning not implemented for this platform
-		ret = 1;
-		// if perror here says 'no such process' and the
-		// process exists still.. is because there's a
-		// missing call to 'wait'. and the process is not
-		// yet available to accept more ptrace queries.
-		if (ret != 0)
-			return false;
-		if (sizeof(regs) < size)
-			size = sizeof(regs);
-		memcpy(buf, &regs, size);
-		return sizeof(regs);
-	} break;
-	}
-	return true;
+	return get_linux_tls_val(dbg, tid);
 }
 
 static int rz_debug_native_reg_read(RzDebug *dbg, int type, ut8 *buf, int size) {
 	if (size < 1) {
 		return false;
 	}
-	return bsd_reg_read(dbg, type, buf, size);
+	return linux_reg_read(dbg, type, buf, size);
 }
 
 static int rz_debug_native_reg_write(RzDebug *dbg, int type, const ut8 *buf, int size) {
 	// XXX use switch or so
 	if (type == RZ_REG_TYPE_DRX) {
-#if __i386__ || __x86_64__
-		return bsd_reg_write(dbg, type, buf, size);
-#else // i386/x86-64
-		return false;
-#endif
+		return linux_reg_write(dbg, type, buf, size);
 	} else if (type == RZ_REG_TYPE_GPR) {
-		return bsd_reg_write(dbg, type, buf, size);
+		return linux_reg_write(dbg, type, buf, size);
 	} else if (type == RZ_REG_TYPE_FPU) {
-		return bsd_reg_write(dbg, type, buf, size);
+		return linux_reg_write(dbg, type, buf, size);
 	} // else eprintf ("TODO: reg_write_non-gpr (%d)\n", type);
 	return false;
 }
 
+static int io_perms_to_prot(int io_perms) {
+	int prot_perms = PROT_NONE;
+
+	if (io_perms & RZ_PERM_R) {
+		prot_perms |= PROT_READ;
+	}
+	if (io_perms & RZ_PERM_W) {
+		prot_perms |= PROT_WRITE;
+	}
+	if (io_perms & RZ_PERM_X) {
+		prot_perms |= PROT_EXEC;
+	}
+	return prot_perms;
+}
+
+static int linux_map_thp(RzDebug *dbg, ut64 addr, int size) {
+	return false;
+}
+
+static RzDebugMap *linux_map_alloc(RzDebug *dbg, ut64 addr, int size, bool thp) {
+	RzBuffer *buf = NULL;
+	RzDebugMap *map = NULL;
+	char code[1024], *sc_name;
+	int num;
+	/* force to usage of x86.as, not yet working x86.nz */
+	char *asm_list[] = {
+		"x86", "x86.as",
+		"x64", "x86.as",
+		NULL
+	};
+
+	/* NOTE: Since kernel 2.4,  that  system  call  has  been  superseded  by
+		 mmap2(2 and  nowadays  the  glibc  mmap()  wrapper  function invokes
+		 mmap2(2)). If arch is x86_32 then usage mmap2() */
+	if (!strcmp(dbg->arch, "x86") && dbg->bits == 4) {
+		sc_name = "mmap2";
+	} else {
+		sc_name = "mmap";
+	}
+	num = rz_syscall_get_num(dbg->analysis->syscall, sc_name);
+#ifndef MAP_ANONYMOUS
+#define MAP_ANONYMOUS 0x20
+#endif
+	snprintf(code, sizeof(code),
+		"sc_mmap@syscall(%d);\n"
+		"main@naked(0) { .rarg0 = sc_mmap(0x%08" PFMT64x ",%d,%d,%d,%d,%d);break;\n"
+		"}\n",
+		num, addr, size, PROT_READ | PROT_WRITE | PROT_EXEC, MAP_ANONYMOUS | MAP_PRIVATE, -1, 0);
+	rz_egg_reset(dbg->egg);
+	rz_egg_setup(dbg->egg, dbg->arch, 8 * dbg->bits, 0, 0);
+	rz_egg_load(dbg->egg, code, 0);
+	if (!rz_egg_compile(dbg->egg)) {
+		eprintf("Cannot compile.\n");
+		goto err_linux_map_alloc;
+	}
+	if (!rz_egg_assemble_asm(dbg->egg, asm_list)) {
+		eprintf("rz_egg_assemble: invalid assembly\n");
+		goto err_linux_map_alloc;
+	}
+	buf = rz_egg_get_bin(dbg->egg);
+	if (buf) {
+		ut64 map_addr;
+
+		rz_reg_arena_push(dbg->reg);
+		ut64 tmpsz;
+		const ut8 *tmp = rz_buf_data(buf, &tmpsz);
+		map_addr = rz_debug_execute(dbg, tmp, tmpsz, 1);
+		rz_reg_arena_pop(dbg->reg);
+		if (map_addr != (ut64)-1) {
+			if (thp) {
+				if (!linux_map_thp(dbg, map_addr, size)) {
+					// Not overly dramatic
+					eprintf("map promotion to huge page failed\n");
+				}
+			}
+			rz_debug_map_sync(dbg);
+			map = rz_debug_map_get(dbg, map_addr);
+		}
+	}
+err_linux_map_alloc:
+	return map;
+}
+
+static int linux_map_dealloc(RzDebug *dbg, ut64 addr, int size) {
+	RzBuffer *buf = NULL;
+	char code[1024];
+	int ret = 0;
+	char *asm_list[] = {
+		"x86", "x86.as",
+		"x64", "x86.as",
+		NULL
+	};
+	int num = rz_syscall_get_num(dbg->analysis->syscall, "munmap");
+
+	snprintf(code, sizeof(code),
+		"sc_munmap@syscall(%d);\n"
+		"main@naked(0) { .rarg0 = sc_munmap(0x%08" PFMT64x ",%d);break;\n"
+		"}\n",
+		num, addr, size);
+	rz_egg_reset(dbg->egg);
+	rz_egg_setup(dbg->egg, dbg->arch, 8 * dbg->bits, 0, 0);
+	rz_egg_load(dbg->egg, code, 0);
+	if (!rz_egg_compile(dbg->egg)) {
+		eprintf("Cannot compile.\n");
+		goto err_linux_map_dealloc;
+	}
+	if (!rz_egg_assemble_asm(dbg->egg, asm_list)) {
+		eprintf("rz_egg_assemble: invalid assembly\n");
+		goto err_linux_map_dealloc;
+	}
+	buf = rz_egg_get_bin(dbg->egg);
+	if (buf) {
+		rz_reg_arena_push(dbg->reg);
+		ut64 tmpsz;
+		const ut8 *tmp = rz_buf_data(buf, &tmpsz);
+		ret = rz_debug_execute(dbg, tmp, tmpsz, 1) == 0;
+		rz_reg_arena_pop(dbg->reg);
+	}
+err_linux_map_dealloc:
+	return ret;
+}
+
 static RzDebugMap *rz_debug_native_map_alloc(RzDebug *dbg, ut64 addr, int size, bool thp) {
-	// malloc not implemented for this platform
-	return NULL;
+	return linux_map_alloc(dbg, addr, size, thp);
 }
 
 static int rz_debug_native_map_dealloc(RzDebug *dbg, ut64 addr, int size) {
-	// mdealloc not implemented for this platform
-	return false;
+	return linux_map_dealloc(dbg, addr, size);
 }
 
 static void _map_free(RzDebugMap *map) {
@@ -269,9 +410,6 @@ static RzList /*<RzDebugMap *>*/ *rz_debug_native_map_get(RzDebug *dbg) {
 	/* prepend 0x prefix */
 	region[0] = region2[0] = '0';
 	region[1] = region2[1] = 'x';
-
-	/* OpenBSD has no procfs, so no idea trying. */
-	return bsd_native_sysctl_map(dbg);
 
 	snprintf(path, sizeof(path), "/proc/%d/maps", dbg->pid);
 
@@ -307,6 +445,7 @@ static RzList /*<RzDebugMap *>*/ *rz_debug_native_map_get(RzDebug *dbg) {
 		if (line_len == 0) {
 			break;
 		}
+
 		ut64 offset = 0;
 		// 7fc8124c4000-7fc81278d000 r--p 00000000 fc:00 17043921 /usr/lib/locale/locale-archive
 		i = sscanf(line, "%" RZ_STR_DEF(PROC_REGION_LEFT_SZ) "s %" RZ_STR_DEF(PROC_PERM_SZ) "s %08" PFMT64x " %*s %*s %" RZ_STR_DEF(PROC_NAME_SZ) "[^\n]", &region[2], perms, &offset, name);
@@ -350,6 +489,8 @@ static RzList /*<RzDebugMap *>*/ *rz_debug_native_map_get(RzDebug *dbg) {
 		if (!map) {
 			break;
 		}
+		map->offset = offset;
+		map->shared = map_is_shared;
 		map->file = strdup(name);
 		rz_list_append(list, map);
 	}
@@ -422,7 +563,6 @@ static bool rz_debug_native_init(RzDebug *dbg, void **user) {
 static void rz_debug_native_fini(RzDebug *dbg, void *user) {
 }
 
-#if __i386__ || __x86_64__
 static void sync_drx_regs(RzDebug *dbg, drxt *regs, size_t num_regs) {
 	/* sanity check, we rely on this assumption */
 	if (num_regs != NUM_DRX_REGISTERS) {
@@ -444,9 +584,7 @@ static void sync_drx_regs(RzDebug *dbg, drxt *regs, size_t num_regs) {
 	regs[6] = rz_reg_getv(R, "dr6");
 	regs[7] = rz_reg_getv(R, "dr7");
 }
-#endif
 
-#if __i386__ || __x86_64__
 static void set_drx_regs(RzDebug *dbg, drxt *regs, size_t num_regs) {
 	/* sanity check, we rely on this assumption */
 	if (num_regs != NUM_DRX_REGISTERS) {
@@ -462,10 +600,8 @@ static void set_drx_regs(RzDebug *dbg, drxt *regs, size_t num_regs) {
 	rz_reg_setv(R, "dr6", regs[6]);
 	rz_reg_setv(R, "dr7", regs[7]);
 }
-#endif
 
 static int rz_debug_native_drx(RzDebug *dbg, int n, ut64 addr, int sz, int rwx, int g, int api_type) {
-#if __i386__ || __x86_64__
 	int retval = false;
 	drxt regs[NUM_DRX_REGISTERS] = { 0 };
 	// sync drx regs
@@ -499,29 +635,55 @@ static int rz_debug_native_drx(RzDebug *dbg, int n, ut64 addr, int sz, int rwx, 
 	set_drx_regs(dbg, regs, NUM_DRX_REGISTERS);
 
 	return retval;
-#else
-	eprintf("drx: Unsupported platform\n");
-#endif
-	return -1;
 }
 
 static int rz_debug_native_bp(RzBreakpoint *bp, RzBreakpointItem *b, bool set) {
 	if (b && b->hw) {
-#if __i386__ || __x86_64__
 		return set
 			? drx_add((RzDebug *)bp->user, bp, b)
 			: drx_del((RzDebug *)bp->user, bp, b);
-#endif
 	}
 	return false;
 }
 
 static RzList /*<RzDebugDesc *>*/ *rz_debug_desc_native_list(int pid) {
-#warning list filedescriptors not supported for this platform
-	return NULL;
+	return linux_desc_list(pid);
 }
 
 static int rz_debug_native_map_protect(RzDebug *dbg, ut64 addr, int size, int perms) {
+	RzBuffer *buf = NULL;
+	char code[1024];
+	int num;
+
+	num = rz_syscall_get_num(dbg->analysis->syscall, "mprotect");
+	snprintf(code, sizeof(code),
+		"sc@syscall(%d);\n"
+		"main@global(0) { sc(%p,%d,%d);\n"
+		":int3\n"
+		"}\n",
+		num, (void *)(size_t)addr, size, io_perms_to_prot(perms));
+
+	rz_egg_reset(dbg->egg);
+	rz_egg_setup(dbg->egg, dbg->arch, 8 * dbg->bits, 0, 0);
+	rz_egg_load(dbg->egg, code, 0);
+	if (!rz_egg_compile(dbg->egg)) {
+		eprintf("Cannot compile.\n");
+		return false;
+	}
+	if (!rz_egg_assemble(dbg->egg)) {
+		eprintf("rz_egg_assemble: invalid assembly\n");
+		return false;
+	}
+	buf = rz_egg_get_bin(dbg->egg);
+	if (buf) {
+		rz_reg_arena_push(dbg->reg);
+		ut64 tmpsz;
+		const ut8 *tmp = rz_buf_data(buf, &tmpsz);
+		rz_debug_execute(dbg, tmp, tmpsz, 1);
+		rz_reg_arena_pop(dbg->reg);
+		return true;
+	}
+
 	return false;
 }
 
@@ -530,5 +692,6 @@ static int rz_debug_desc_native_open(const char *path) {
 }
 
 static bool rz_debug_gcore(RzDebug *dbg, char *path, RzBuffer *dest) {
+	(void)path;
 	return false;
 }
