+++ librz/egg/emit_x86.c
@@ -170,7 +170,7 @@ static void emit_string(RzEgg *egg, const char *dstvar, const char *str, int j) 
 
 	/* XXX: Hack: Adjust offset in RZ_BP correctly for 64b addresses */
 #define BPOFF  (RZ_SZ - 4)
-#define M32(x) (unsigned int)((x)&0xffffffff)
+#define M32(x) (unsigned int)((x) & 0xffffffff)
 	/* XXX: Assumes sizeof(ut32) == 4 */
 	for (i = 4; i <= oj; i += 4) {
 		/* XXX endian issues (non-portable asm) */
